+++ plugins/cli/plugin.go
@@ -1,7 +1,6 @@
 package cli
 
 import (
-	"flag"
 	"fmt"
 	"strings"
 	"time"
@@ -21,7 +20,7 @@ import (
 
 var (
 	// AppVersion version number
-	AppVersion          = "0.5.3"
+	AppVersion          = "0.5.4"
 	LatestGithubVersion = AppVersion
 
 	// AppName app code name
@@ -45,15 +44,13 @@ func init() {
 	}
 
 	node.Events.AddPlugin.Attach(events.NewClosure(onAddPlugin))
-
-	flag.Usage = printUsage
 }
 
 func parseParameters() {
-	for _, pluginName := range config.NodeConfig.GetStringSlice(node.CFG_DISABLE_PLUGINS) {
+	for _, pluginName := range config.NodeConfig.GetStringSlice(config.CfgNodeDisablePlugins) {
 		node.DisabledPlugins[strings.ToLower(pluginName)] = true
 	}
-	for _, pluginName := range config.NodeConfig.GetStringSlice(node.CFG_ENABLE_PLUGINS) {
+	for _, pluginName := range config.NodeConfig.GetStringSlice(config.CfgNodeEnablePlugins) {
 		node.EnabledPlugins[strings.ToLower(pluginName)] = true
 	}
 }