+++ plugins/tangle/tangle_processor.go
@@ -3,6 +3,7 @@ package tangle
 import (
 	"fmt"
 	"runtime"
+	"sync"
 
 	"github.com/iotaledger/hive.go/daemon"
 	"github.com/iotaledger/hive.go/events"
@@ -28,6 +29,8 @@ var (
 	lastIncomingTPS uint32
 	lastNewTPS      uint32
 	lastOutgoingTPS uint32
+
+	startWaitGroup sync.WaitGroup
 )
 
 func configureTangleProcessor(_ *node.Plugin) {
@@ -50,6 +53,8 @@ func configureTangleProcessor(_ *node.Plugin) {
 
 func runTangleProcessor(_ *node.Plugin) {
 	log.Info("Starting TangleProcessor ...")
+
+	startWaitGroup.Add(4)
 
 	onTransactionProcessed := events.NewClosure(func(transaction *hornet.Transaction, request *rqueue.Request, p *peer.Peer) {
 		receiveTxWorkerPool.Submit(transaction, request, p)
@@ -75,6 +80,7 @@ func runTangleProcessor(_ *node.Plugin) {
 
 	daemon.BackgroundWorker("TangleProcessor[UpdateMetrics]", func(shutdownSignal <-chan struct{}) {
 		metricsplugin.Events.TPSMetricsUpdated.Attach(onTPSMetricsUpdated)
+		startWaitGroup.Done()
 		<-shutdownSignal
 		metricsplugin.Events.TPSMetricsUpdated.Detach(onTPSMetricsUpdated)
 	}, shutdown.PriorityMetricsUpdater)
@@ -83,6 +89,7 @@ func runTangleProcessor(_ *node.Plugin) {
 		log.Info("Starting TangleProcessor[ReceiveTx] ... done")
 		gossip.Processor().Events.TransactionProcessed.Attach(onTransactionProcessed)
 		receiveTxWorkerPool.Start()
+		startWaitGroup.Done()
 		<-shutdownSignal
 		log.Info("Stopping TangleProcessor[ReceiveTx] ...")
 		gossip.Processor().Events.TransactionProcessed.Detach(onTransactionProcessed)
@@ -95,6 +102,7 @@ func runTangleProcessor(_ *node.Plugin) {
 		processValidMilestoneWorkerPool.Start()
 		tangle.Events.ReceivedValidMilestone.Attach(onReceivedValidMilestone)
 		tangle.Events.ReceivedInvalidMilestone.Attach(onReceivedInvalidMilestone)
+		startWaitGroup.Done()
 		<-shutdownSignal
 		log.Info("Stopping TangleProcessor[ProcessMilestone] ...")
 		tangle.Events.ReceivedValidMilestone.Detach(onReceivedValidMilestone)
@@ -106,11 +114,18 @@ func runTangleProcessor(_ *node.Plugin) {
 	daemon.BackgroundWorker("TangleProcessor[MilestoneSolidifier]", func(shutdownSignal <-chan struct{}) {
 		log.Info("Starting TangleProcessor[MilestoneSolidifier] ... done")
 		milestoneSolidifierWorkerPool.Start()
+		startWaitGroup.Done()
 		<-shutdownSignal
 		log.Info("Stopping TangleProcessor[MilestoneSolidifier] ...")
 		milestoneSolidifierWorkerPool.StopAndWait()
 		log.Info("Stopping TangleProcessor[MilestoneSolidifier] ... done")
 	}, shutdown.PriorityMilestoneSolidifier)
+
+}
+
+// WaitForTangleProcessorStartup waits until all background workers of the tangle processor are started.
+func WaitForTangleProcessorStartup() {
+	startWaitGroup.Wait()
 }
 
 func IsReceiveTxWorkerPoolBusy() bool {