+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.5.4] - 13.11.2020
+
+### Changed
+
+    - use iota.go/curl and iota.go/curl/bct instead of the much less performant hive.go versions
+    - migrate batchhasher from hive.go into hornet
+    - optimize batchhasher and add testcases
+    - iota.go now ignores 243rd trit in signature fragments
+
+### Fixed
+
+    - coordinator startup race condition
+    - improve tryte distribution of tools/rand-seed
+    - do not change default GOMAXPROCS value
+    - several CLI flag problems
+    - delete invalid milestones (2272660, 2272661) that could not be solidified after COO crashed
+
 ## [0.5.3] - 14.09.2020
 
 ### Added